Implement Wayland support
Implement the wlr_gamma_control_unstable_v1 protocol.
The Wayland backend is prioritized over randr since it will fail in
an X11 context anyway (but the randr backend would not fail in a Wayland
context with xwayland support).
The gamma setting is reset when the client disconnects, thus the
behavior for oneshot mode blocks, like for Quartz.

Co-Authored-By: Giulio Camuffo <giuliocamuffo@gmail.com>
Co-Authored-By: Thomas Weißschuh <thomas@t-8ch.de>
Co-Authored-By: Marek Otahal <markotahal@gmail.com>
Co-Authored-By: Dominique Martinet <asmadeus@codewreck.org>